Move translateLMIndexVarsToProjections and ExtractNonConstPartsOfIndexCondition
