+++ arangod/Aql/ExecutionNode/IndexNode.cpp
@@ -33,7 +33,7 @@
 #include "Aql/Executor/IndexExecutor.h"
 #include "Aql/Expression.h"
 #include "Aql/NonConstExpressionContainer.h"
-#include "Aql/OptimizerUtils.h"
+#include "Aql/Optimizer/Utils/ExtractNonConstPartsOfIndexCondition.h"
 #include "Aql/Projections.h"
 #include "Aql/Query.h"
 #include "Aql/RegisterPlan.h"
@@ -44,12 +44,53 @@
 #include "StorageEngine/StorageEngine.h"
 #include "Transaction/CountCache.h"
 #include "Transaction/Methods.h"
+#include "VocBase/LogicalCollection.h"
 
 #include <velocypack/Iterator.h>
 #include <cstdint>
 
-using namespace arangodb;
-using namespace arangodb::aql;
+namespace arangodb::aql {
+
+namespace {
+Projections translateLMIndexVarsToProjections(
+    ExecutionPlan* plan, IndexNode::IndexValuesVars const& indexVars,
+    transaction::Methods::IndexHandle index) {
+  // Translate the late materialize "projections" description
+  // into the usual projections description
+  auto& coveredFields = index->coveredFields();
+
+  std::vector<AttributeNamePath> projectedAttributes;
+  for (auto [var, fieldIndex] : indexVars.second) {
+    auto& field = coveredFields[fieldIndex];
+    std::vector<std::string> fieldCopy;
+    fieldCopy.reserve(field.size());
+    std::transform(field.begin(), field.end(), std::back_inserter(fieldCopy),
+                   [&](auto const& attr) {
+                     TRI_ASSERT(attr.shouldExpand == false);
+                     return attr.name;
+                   });
+    projectedAttributes.emplace_back(std::move(fieldCopy),
+                                     plan->getAst()->query().resourceMonitor());
+  }
+
+  Projections projections{std::move(projectedAttributes)};
+
+  std::size_t i = 0;
+  for (auto [var, fieldIndex] : indexVars.second) {
+    auto& proj = projections[i++];
+    proj.coveringIndexPosition = fieldIndex;
+    proj.coveringIndexCutoff = proj.path.size();
+    proj.variable = var;
+    proj.levelsToClose = proj.startsAtLevel = 0;
+    proj.type = proj.path.type();
+  }
+
+  if (index->covers(projections)) {
+    projections.setCoveringContext(index->collection().id(), index);
+  }
+  return projections;
+}
+}  // namespace
 
 IndexNode::IndexNode(
     ExecutionPlan* plan, ExecutionNodeId id, Collection const* collection,
@@ -186,8 +227,8 @@ IndexNode::IndexNode(ExecutionPlan* plan,
     }
 
     // convert old index value vars to projections
-    utils::translateLMIndexVarsToProjections(plan, oldIndexValuesVars,
-                                             getSingleIndex());
+    translateLMIndexVarsToProjections(plan, oldIndexValuesVars,
+                                      getSingleIndex());
   }
   _options.forLateMaterialization = isLateMaterialized();
 
@@ -318,7 +359,7 @@ NonConstExpressionContainer IndexNode::buildNonConstExpressions() const {
                                      "The index id cannot be empty.");
     }
 
-    return utils::extractNonConstPartsOfIndexCondition(
+    return optimizer::extractNonConstPartsOfIndexCondition(
         _plan->getAst(), getRegisterPlan()->varInfo, options().evaluateFCalls,
         idx.get(), _condition->root(), _outVariable);
   }
@@ -724,8 +765,8 @@ void IndexNode::setLateMaterialized(aql::Variable const* docIdVariable,
                                           indVars.second.indexFieldNum);
   }
 
-  _projections = utils::translateLMIndexVarsToProjections(
-      plan(), oldIndexValuesVars, getSingleIndex());
+  _projections = translateLMIndexVarsToProjections(plan(), oldIndexValuesVars,
+                                                   getSingleIndex());
   _options.forLateMaterialization = true;
 }
 
@@ -824,3 +865,5 @@ aql::Variable const* IndexNode::getLateMaterializedDocIdOutVar() const {
   TRI_ASSERT(isLateMaterialized());
   return _outNonMaterializedDocId;
 }
+
+}  // namespace arangodb::aql
