+++ arangod/Aql/ExecutionNode/TraversalNode.cpp
@@ -33,6 +33,7 @@
 #include "Aql/Executor/TraversalExecutor.h"
 #include "Aql/Expression.h"
 #include "Aql/OptimizerUtils.h"
+#include "Aql/Optimizer/Utils/ExtractNonConstPartsOfIndexCondition.h"
 #include "Aql/PruneExpressionEvaluator.h"
 #include "Aql/RegisterPlan.h"
 #include "Aql/SingleRowFetcher.h"
@@ -812,7 +813,7 @@ std::vector<IndexAccessor> TraversalNode::buildIndexAccessor(
     std::unique_ptr<aql::Expression> expression =
         generateExpression(remainderCondition, indexCondition);
 
-    auto container = aql::utils::extractNonConstPartsOfIndexCondition(
+    auto container = aql::optimizer::extractNonConstPartsOfIndexCondition(
         ast, getRegisterPlan()->varInfo, false, nullptr, indexCondition,
         options()->tmpVar());
     indexAccessors.emplace_back(std::move(indexToUse), indexCondition,
