+++ arangod/Aql/Optimizer/Utils/CMakeLists.txt
@@ -1,7 +1,8 @@
 target_sources(arango_aql PRIVATE
   AccessesCollectionVariable.cpp
-  FindProjections.cpp
   EnumeratePathsFilterMatcher.cpp
+  ExtractNonConstPartsOfIndexCondition.cpp
+  FindProjections.cpp
   LateMaterializedCommon.cpp
   OptimizeTraversalPathVariable.cpp
 )
